new style for side menu

+++ db/migrate/20170916230747_create_works.rb
@@ -2,7 +2,7 @@ class CreateWorks < ActiveRecord::Migration[5.0]
   def change
     create_table :works do |t|
       t.date :date
-      t.integer :hour
+      t.float :hour
       t.string :description
       t.references :project, foreign_key: true
       t.references :time_sheet, foreign_key: true
